==> 00-mk_pickle.ipynb <==
import numpy as np
import pandas as pd
import os
import csv
import re

First download all labelled samples per mainshock rupture model, in csv format, into AllCSV folder from the DeVries18 Google Drive from [this](https://drive.google.com/drive/folders/1c5Rb_6EsuP2XedDjg37bFDyf8AadtGDa) link.

This link is also given [here](https://github.com/phoebemrdevries/Learning-aftershock-location-patterns)

directory = 'AllCSV/'
file_list = os.listdir(directory)
n_files = len(file_list)
file_list.sort()

### Read an template file to get the name of columns    

file1 = 'AllCSV/1968HYUGAx01YAGI_grid.csv'
file1_tmp = pd.read_csv(file1)
file_columns = list(file1_tmp.columns)
file_columns

indata = []
for i in range(len(file_list)):
    mainshock_name = 's' + file_list[i][0:file_list[i].find('_')]
    df_tmp = pd.read_csv(directory + file_list[i])
    df_tmp['ID'] = np.repeat(mainshock_name, len(df_tmp))
    indata.append(df_tmp)

df = pd.concat(indata)
len(df)

df

df.to_pickle('./Data/complete_dataset_labelled.pkl')
