==> 04-pred_seism_aftXYZ.model_def.ipynb <==
# Model Definition

Model definition step for the project '**Aftershock pattern prediction based on earthquake rupture data for improved seismic hazard assessment**' DeVries18 will refer to the article 'Deep learning of aftershock patterns following large earthquakes' by Phoebe M. R. DeVries, Fernanda Viégas, Martin Wattenberg & Brendan J. Meade, and published in Nature in 2018 (https://www.nature.com/articles/s41586-018-0438-y ).

from keras.models import Sequential
from keras.layers import Dense, Activation, Dropout
from keras import optimizers
from keras import initializers
from keras import metrics

## Define baseline model (DeVries18)

We here reproduce the DeVries18 deep learning model = six hidden layers with 50 neurons each and hyperbolic tangent activation functions (13,451 weights and biases in total).

# same model as in https://github.com/phoebemrdevries/Learning-aftershock-location-patterns

def create_baselinemodel_DeVries18():
    model = Sequential()
    model.add(Dense(50, input_dim=12, kernel_initializer='lecun_uniform', activation = 'tanh'))
    model.add(Dropout(0.50))
    model.add(Dense(50, kernel_initializer='lecun_uniform', activation= 'tanh'))
    model.add(Dropout(0.50))
    model.add(Dense(50, kernel_initializer='lecun_uniform', activation= 'tanh'))
    model.add(Dropout(0.50))
    model.add(Dense(50, kernel_initializer='lecun_uniform', activation= 'tanh'))
    model.add(Dropout(0.50))
    model.add(Dense(50, kernel_initializer='lecun_uniform', activation= 'tanh'))
    model.add(Dropout(0.50))
    model.add(Dense(50, kernel_initializer='lecun_uniform', activation= 'tanh'))
    model.add(Dropout(0.50))
    model.add(Dense(1, kernel_initializer='lecun_uniform', activation='sigmoid'))
    model.compile(optimizer='adadelta', loss='binary_crossentropy', metrics=[metrics.binary_accuracy])
    return model

create_baselinemodel_DeVries18().save('./Models/model_baseline_prev.h5')

## Define new models

### Define simplified DNN for DeVries18 features

Since the DeVries model has a 12-neuron input layer, we first verify whether a much simpler Deep Neural Network (DNN) topology can lead to similar results. We also reduce the dropout rate. We keep everything else the same (i.e. kernel_initializer, activation, optimizer, loss). This model is for illustration purposes only. A different model with different features is proposed in this project (see below).

def create_baselinemodel_DeVries18_simplified():
    model = Sequential()
    model.add(Dense(8, input_dim=12, kernel_initializer='lecun_uniform', activation = 'tanh'))
    model.add(Dropout(0.20))
    model.add(Dense(8, kernel_initializer='lecun_uniform', activation= 'tanh'))
    model.add(Dropout(0.20))
    model.add(Dense(1, kernel_initializer='lecun_uniform', activation='sigmoid'))
    model.compile(optimizer='adadelta', loss='binary_crossentropy', metrics=[metrics.binary_accuracy])
    return model

create_baselinemodel_DeVries18_simplified().save('./Models/model_baseline_DeVries18_simplified_init.h5')

### Define new DNN for new features

We use more direct features than in DeVries18, here based on geometry/kinematics instead of stress. We further simplify the topology of the DNN compared to the baseline.

from keras.models import Sequential
from keras.layers import Dense, Activation, Dropout
from keras import optimizers
from keras import initializers
from keras import metrics

def create_model_DNN():
    model = Sequential()
    model.add(Dense(6, input_dim=2, kernel_initializer='lecun_uniform', activation = 'relu'))
    model.add(Dropout(0.20))
    model.add(Dense(6, kernel_initializer='lecun_uniform', activation= 'relu'))
    model.add(Dropout(0.20))
    model.add(Dense(1, kernel_initializer='lecun_uniform', activation='sigmoid'))
    model.compile(optimizer='adadelta', loss='binary_crossentropy', metrics=[metrics.binary_accuracy])
    return model

# metrics = binary accuracy as in the baseline model.
# AUC not being supported by Keras, we will only use it during final model evaluation (as presented also in DeVries18).
# Binary accuracy is a common metric for binary classification. However the Area Under the Receiver-Operating Characteristic (ROC) Curve (AUC)
# will provide more information than the accuracy by considering all possible classification threshold values.

create_model_DNN().save('./Models/model_DNN_init.h5')

### Define ANN for new features

We also test a shallow Artificial Neural Network (ANN) (i.e. no deep learning, only one hidden layer).

NB: During training, we will also test another standard machine learning algorithm (XGBoost) for completeness.

def create_model_ANN():
    model = Sequential()
    model.add(Dense(30, input_dim=2, kernel_initializer='lecun_uniform', activation = 'relu'))
    model.add(Dense(1, kernel_initializer='lecun_uniform', activation='sigmoid'))
    model.compile(optimizer='adadelta', loss='binary_crossentropy', metrics=[metrics.binary_accuracy])
    return model

create_model_ANN().save('./Models/model_ANN_init.h5')

!ls ./Models
