==> 05-pred_seism_aftXYZ.model_train.ipynb <==
# Model Training

Model training step for the project '**Aftershock pattern prediction based on earthquake rupture data for improved seismic hazard assessment**' (pred_seism_aftXYZ). DeVries18 will refer to the article 'Deep learning of aftershock patterns following large earthquakes' by Phoebe M. R. DeVries, Fernanda Viégas, Martin Wattenberg & Brendan J. Meade, and published in Nature in 2018 (https://www.nature.com/articles/s41586-018-0438-y ).

Inputs from previous steps of the process model:
-  'Features.pkl': Features and target of baseline model (DeVries18), as pickle file;
-  'new_features.pkl': New features and same target as baseline, as pickle file;
-  'model_baseline_prev.h5': Baseline deep neural net of DeVries18, as (untrained) Keras model HDF5 file;
-  'model_baseline_DeVries18_simplified_init.h5': Simplified topology for baseline DNN of DeVries18, as (untrained) Keras model HDF5 file;
-  'model_DNN_init.h5': Deep neural net topology proposed for this project, as (untrained) Keras model HDF5 file;
-  'model_ANN_init.h5': Shallow artificial neural net proposed for this project, as (untrained) Keras model HDF5 file.

For comparison with the DeVries study, we will use the same training and test sets ('Training_FileNames.h5', 'Testing_FileNames.h5'), first imported from the [Google Drive](https://drive.google.com/drive/folders/1c5Rb_6EsuP2XedDjg37bFDyf8AadtGDa)

## Split dataset into training and testing sets

import h5py
import numpy as np
import pandas as pd
import sklearn

h5file1 = h5py.File('./Data/Training_FileNames.h5', 'r')
training_filenames = np.array(h5file1.get('file_names_training'))
h5file2 = h5py.File('./Data/Testing_FileNames.h5', 'r')
testing_filenames = np.array(h5file2.get('file_names_testing'))

training_IDs_temp = map(lambda x: str(x, 'utf-8'), training_filenames)  
training_IDs = list(map(lambda x: 's' + x[0:x.find('_')], training_IDs_temp))
testing_IDs_temp = map(lambda x: str(x, 'utf-8'), testing_filenames)   
testing_IDs = list(map(lambda x: 's' + x[0:x.find('_')], testing_IDs_temp))

Features = pd.read_pickle('./Data/Features.pkl')
new_features = pd.read_pickle('./Data/new_features.pkl')

# create training & testing sets
TrainingSet_prev = Features.loc[Features['ID'].isin(training_IDs)]
TestingSet_prev = Features.loc[Features['ID'].isin(testing_IDs)]
TrainingSet_new = new_features.loc[new_features['ID'].isin(training_IDs)]
TestingSet_new = new_features.loc[new_features['ID'].isin(testing_IDs)]

# ratio of training samples
[len(TrainingSet_prev), len(TestingSet_prev), len(TrainingSet_prev)/(len(TrainingSet_prev)+len(TestingSet_prev))]

# shuffle training sets
TrainingSet_prev = sklearn.utils.shuffle(TrainingSet_prev)
TrainingSet_new = sklearn.utils.shuffle(TrainingSet_new)

## Train baseline model

TrainingSet_prev.head(10)

features = ['posabsxx', 'posabsxy', 'posabsyy', 'posabsxz', 'posabsyz', 'posabszz',
           'negabsxx', 'negabsxy', 'negabsyy', 'negabsxz', 'negabsyz', 'negabszz']

target = 'aftershocksyn'

x_train = TrainingSet_prev[features]
y_train = TrainingSet_prev[target]
x_test = TestingSet_prev[features]
y_test = TestingSet_prev[target]

x_test.to_pickle("./Data/TestingSet_X_prev.pkl")
y_test.to_pickle("./Data/TestingSet_y_prev.pkl")

import keras
from keras.models import load_model

# same as in DeVries18
baselinemodel_prev = load_model('./Models/model_baseline_prev.h5')
batch_size = 3500
epochs = 5

history = baselinemodel_prev.fit(x_train, y_train,
                            batch_size = batch_size,
                            epochs = epochs,
                            validation_split = 0.1)

This is the baseline model CV results - we should aim at achieving similar or greater accuracy ($\geq$ 72%) for the new model with new set of features. This will however not assure a same or better generalization than the baseline model (AUC = 85%, see next step of the process model).

baselinemodel_prev.save('./Models/model_baseline_prev_trained.h5')

### Side note

Since the Devries18 model uses a relatively small input layer (12 nodes), a simpler DNN topology should do just fine. We show it, only for illustrative purpose. The model proposed in this project will use another set of features.

baselinemodel_DeVries18_simplified = load_model('./Models/model_baseline_DeVries18_simplified_init.h5')

history = baselinemodel_DeVries18_simplified.fit(x_train, y_train,
                            batch_size = batch_size,
                            epochs = epochs,
                            validation_split = 0.1)

# note the similar CV accuracy using 12 - 8 - 8 - 1 instead of 12 - 50 - 50 - 50 - 50 - 50 - 50 - 1

baselinemodel_DeVries18_simplified.save('./Models/model_baseline_DeVries18_simplified_trained.h5')

## Train new models

Training new models with the proposed set of features based on geometry and kinematics (minimum distance to mainshock rupture and mean slip on rupture).

#features = ['mindist', 'dipMean', 'strikeMean', 'slipMean']   #best 4 - 12 - 12 - 1
features = ['mindist', 'slipMean']   #best 2 - 6 - 6 - 1
target = 'aftershocksyn'

x_train = TrainingSet_new[features]
y_train = TrainingSet_new[target]
x_test = TestingSet_new[features]
y_test = TestingSet_new[target]

# save for next step of the process model
x_test.to_pickle("./Data/TestingSet_X_new.pkl")
y_test.to_pickle("./Data/TestingSet_y_new.pkl")

### Train new DNN (simplified topology)

model_DNN = load_model('./Models/model_DNN_init.h5')

history = model_DNN.fit(x_train, y_train,
                            batch_size = batch_size,
                            epochs = epochs,
                            validation_split = 0.1)

# Documenting a first DNN model result
# we tried the following topology 4 - 12 - 12 - 1, the 4 features being: ['mindist', 'dipMean', 'strikeMean', 'slipMean']
# we obtained: 
#Epoch 5/5
#147801/147801 [==============================] - 0s 2us/step - loss: 0.6134 - binary_accuracy: 0.6829 - 
#                                                           val_loss: 0.5801 - val_binary_accuracy: 0.7228

# additional tuning involved changes in topology, using relu instead of tanh activation, dropout rate change [NOT SHOWN for clarity].

model_DNN.save('./Models/model_DNN_trained.h5')

### Train ANN

Training of an Artificial Neural Network with one hidden layer (no deep learning!).

model_ANN = load_model('./Models/model_ANN_init.h5')

history = model_ANN.fit(x_train, y_train,
                            batch_size = batch_size,
                            epochs = epochs,
                            validation_split = 0.1)

model_ANN.save('./Models/model_ANN_trained.h5')

### Train XGBoost classifier

To not only test neural networks, we here test another standard machine learning algorithm. XGBoost has become a method of choice in recent Kaggle competitions (althought this does not mean that it will perform better than neural networks in the present case - "No Free Lunch" theorem):

import xgboost as xgb
from sklearn.model_selection import GridSearchCV

params = {
    'max_depth': [5, 7, 10]
}

gridsearch = GridSearchCV(estimator = xgb.XGBClassifier(
                          objective = "binary:logistic"
                          ),
                        param_grid = params,
                        scoring='accuracy',
                        n_jobs=1,
                        cv=5)

gridsearch.fit(x_train, y_train)

gridsearch.cv_results_, gridsearch.best_params_, gridsearch.best_score_

from joblib import dump

dump(gridsearch, './Models/model_XGBoost.joblib')
